Write temp files into temp directory for sorted table

+++ presto-hive/src/main/java/com/facebook/presto/hive/SortingFileWriter.java
@@ -73,6 +73,7 @@ public class SortingFileWriter
     private final HiveFileWriter outputWriter;
     private final SortBuffer sortBuffer;
     private final TempFileSinkFactory tempFileSinkFactory;
+    private final boolean sortedWriteToTempPathEnabled;
     private final Queue<TempFile> tempFiles = new PriorityQueue<>(comparing(TempFile::getSize));
 
     public SortingFileWriter(
@@ -85,7 +86,8 @@ public class SortingFileWriter
             List<Integer> sortFields,
             List<SortOrder> sortOrders,
             PageSorter pageSorter,
-            TempFileSinkFactory tempFileSinkFactory)
+            TempFileSinkFactory tempFileSinkFactory,
+            boolean sortedWriteToTempPathEnabled)
     {
         checkArgument(maxOpenTempFiles >= 2, "maxOpenTempFiles must be at least two");
         this.fileSystem = requireNonNull(fileSystem, "fileSystem is null");
@@ -97,6 +99,7 @@ public class SortingFileWriter
         this.outputWriter = requireNonNull(outputWriter, "outputWriter is null");
         this.sortBuffer = new SortBuffer(maxMemory, types, sortFields, sortOrders, pageSorter);
         this.tempFileSinkFactory = tempFileSinkFactory;
+        this.sortedWriteToTempPathEnabled = sortedWriteToTempPathEnabled;
     }
 
     @Override
@@ -146,8 +149,10 @@ public class SortingFileWriter
     @Override
     public void rollback()
     {
-        for (TempFile file : tempFiles) {
-            cleanupFile(file.getPath());
+        if (!sortedWriteToTempPathEnabled) {
+            for (TempFile file : tempFiles) {
+                cleanupFile(file.getPath());
+            }
         }
 
         outputWriter.rollback();
@@ -223,11 +228,13 @@ public class SortingFileWriter
             new MergingPageIterator(iterators, types, sortFields, sortOrders)
                     .forEachRemaining(consumer);
 
-            for (TempFile tempFile : files) {
-                Path file = tempFile.getPath();
-                fileSystem.delete(file, false);
-                if (fileSystem.exists(file)) {
-                    throw new IOException("Failed to delete temporary file: " + file);
+            if (!sortedWriteToTempPathEnabled) {
+                for (TempFile tempFile : files) {
+                    Path file = tempFile.getPath();
+                    fileSystem.delete(file, false);
+                    if (fileSystem.exists(file)) {
+                        throw new IOException("Failed to delete temporary file: " + file);
+                    }
                 }
             }
         }
@@ -246,7 +253,9 @@ public class SortingFileWriter
             tempFiles.add(new TempFile(tempFile, writer.getWrittenBytes()));
         }
         catch (IOException | UncheckedIOException e) {
-            cleanupFile(tempFile);
+            if (!sortedWriteToTempPathEnabled) {
+                cleanupFile(tempFile);
+            }
             throw new PrestoException(HIVE_WRITER_DATA_ERROR, "Failed to write temporary file: " + tempFile, e);
         }
     }
